backend/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
Pillow==8.0.1
PyJWT==1.7.1
pytz==2020.4
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
